Updated docker and created automatic database creation

+++ docker/production/build.sh
@@ -1,6 +1,22 @@
-rm -r client/src
-rm -r server/src
+#!/bin/bash
 
+function removeFolders {
+    if test -e client/src; then 
+        rm -r client/src
+    fi
+
+    if test -e server/src; then
+        rm -r server/src
+    fi
+}
+
+GREEN='\033[0;32m'
+NC='\033[0m'
+
+printf "${GREEN}Try to remove olf files in the docker container\n\n${NC}"
+removeFolders
+
+printf "${GREEN}Copy src files in the docker folder\n\n${NC}"
 mkdir client/src
 mkdir server/src
 
@@ -13,7 +29,8 @@ cp -r ../../server/package*.json server/src
 cp -r ../../server/src server/src
 cp -r ../../server/tsconfig*.json server/src
 
+
+printf "${GREEN}Build docker compose\n\n${NC}"
 docker-compose build --no-cache
 
-rm -r client/src
-rm -r server/src
+removeFolders
